requirements.txt
aliyun-python-sdk-core==2.13.4
aliyun-python-sdk-core-v3==2.13.3
aliyun-python-sdk-kms==2.6.0
amqp==2.4.2
billiard==3.5.0.5
celery
certifi==2019.3.9
chardet==3.0.4
colorlog==3.1.4
crcmod==1.7
Django==2.1
django-aliyun-oss2==0.1.1
django-cors-headers==1.2.2
django-extensions==1.9.8
django-filter==2.0.0
django-oauth-toolkit==1.2.0
django-ranged-response==0.2.0
django-redis==4.10.0
django-simple-captcha==0.5.11
djangorestframework==3.9.3
gevent
greenlet==0.4.15
gunicorn
idna==2.8
jmespath==0.9.4
kombu==4.5.0
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
oauthlib==3.0.1
oss2==2.6.1
Pillow==6.0.0
protobuf==3.7.1
pycryptodome==3.8.1
PyMySQL==0.9.3
pytz==2019.1
redis==3.2.1
requests==2.21.0
rest-condition==1.0.3
six==1.12.0
sqlparse==0.3.0
typing==3.6.6
urllib3==1.24.2
vine==1.3.0
